feat(environ)!: add environ module and move inside it Mapping.Environment class
